ユースケースに漏れていたchat gpt部分をserviceに切り出した。
